post and user seeded

+++ server/config/seedDB.js
@@ -14,18 +14,34 @@ const DB_NAME = process.env.DB_NAME;
         mongoose.connect(`${MONGO_URL}/${DB_NAME}`)
         .then(() => console.log(`Connected for seeding: ${DB_NAME}`));
 
+        const userCount = await User.find().countDocuments();
+        if(userCount > 10) return console.log(`Records are enough: DB hit > 10 users`);
+
         const customUser = await User.findOne({username:'ricky'});
         // erasing dummy data
         const namequery = { username: {$ne: customUser.username} };
         const count = await User.find(namequery).countDocuments();
         await User.deleteMany(namequery);
-        console.log(`${count} users deleted`);
+        console.log(`${count} Users deleted`);
+
 
         await Post.deleteMany({ author: { $ne: customUser._id } });
+        console.log(`Posts Deleted`)
 
+
+        // seeding users
         const users = seed.genereateDummyUsers();
         const seededUsers = await User.insertMany(users)
-        console.log(`Dummy Users:: Number: ${seededUsers.length}`, seededUsers);
+        console.log(`User seeded`)
+        // console.log(`Dummy Users:: Number: ${seededUsers.length}`, seededUsers);
+
+
+        // seeding posts
+        const posts = seed.generateDummyPosts(seededUsers);
+        const seededPosts = await Post.insertMany(posts);
+        console.log(`Post Seeded`);
+        // console.log(`seeded posts`, seededPosts)
+
 
     mongoose.connection.close();
